+++ src/interfaces/IMetadata.ts
@@ -1,4 +1,5 @@
 import { Claims, RouteAction, Scopes } from '../enums';
+import { IMutex } from './IMutex';
 import { IRequest, IResponse } from './IStore';
 
 /**
@@ -15,15 +16,13 @@ export interface IClassMetadata {
  * The optional properties align with what the decorators might store.
  */
 export interface IMethodMetadata {
-  automaticRefresh?: boolean;
   requiresAuth?: boolean;
-  requiredClaim?: {
-    claimKey: string;
-    claimValue?: any;
-  };
-  requiredScopes?: Scopes[];
+  requiredClaims?: Claims[];
   isOidcCallback?: boolean;
   isOidcLogin?: boolean;
+  isWithMutexLock?: boolean;
+  mutex?: IMutex;
+  timeout?: number;
 }
 
 /**